HLS playlist load timer compliant with Pantos 6.3.4
Changes the load timer calculation to:

1. Reload, if playlist changed, per last segment duration, ignoring TARGETDURATION
2. If no change, still 1/2 TARGETDURATION
3. Removes the download time from the next poll time (so time is request to request, not end of last request to next request)

Per the Nov 2022, HLS spec:

[Section 6.3.4. Reloading the Media Playlist File](https://datatracker.ietf.org/doc/html/draft-pantos-hls-rfc8216bis-12#section-6.3.4)

Changes 1 and 3:

>  When a client loads a Playlist file for the first time or reloads a
>  Playlist file and finds that it has changed since the last time it
>  was loaded, the client MUST wait for at least the duration of the
>  last segment in the Playlist before attempting to reload the Playlist
>  file again, measured from the last time the client began loading the
>  Playlist file.

This matters for transcoding Drop Frame (29.97FPS) where the GOP size is non-integer and EXT-X-TARGETDURATION must be an integer.

So, with a GOP segment of 2.002 seconds, a 3 GOP segment is 6.006 seconds duration.  Following the
rule in section *4.4.3.1. EXT-X-TARGETDURATION* yields a 7 second target duration.
This puts the normal (change detected) polling nearly a second behind.