Measures frame rate for iFrame correctly for VOD
The issue was the `FrameRateAnalyzer` was not being called for VOD playlist because
the `onTimelineChanged()` event only fires once for VOD and only for the initial primary
playlist.  The `Timeline` and it's `Window` creating happen only for the primary playlist
when playback starts.   Trickplay switches to an iFrame only playlist so the load of the
iFrame only playlist does not trigger an event.

These bugs are fixed by the change:

* [WSIPCL-15602](https://jira.xperi.com/browse/WSIPCL-15602)
* [WSIPCL-15683](https://jira.xperi.com/browse/WSIPCL-15683)

This fix moves the `FrameRateAnalzyer` to the `DualModeHlsParserFactory` rather
then creating it in the `TrickPlayController` and monitoring timeline change.   This
way the frame rate analyze is performed simply on every playlist load.

For this to work properly with ExoPlayerPlayer a factory method is exposed,

To use this, remove this code:

```java
    	Boolean dualModeProp = getBooleanProperty("disableDualMode");
...
        	factory.setPlaylistParserFactory(new DualModeHlsPlaylistParserFactory(new DefaultHlsPlaylistParserFactory()));
        }
```

And add this code

```java
    	boolean enableDualMode = ...
    gTrickPlayControlFactory.createTrickPlayControl(gExoTrackSelector);
    if (mediaSourceFactory instanceof HlsMediaSource.Factory) {
    	HlsMediaSource.Factory factory = ((HlsMediaSource.Factory) mediaSourceFactory);
    	factory.setPlaylistParserFactory(gTrickPlayControlFactory.createHlsPlaylistParserFactory(enableDualMode))
    }

```