Buffered duration includes only loaded samples
The decision in `shouldContinueLoading()` was using `getNextLoadPosition()` for computing the buffered duration to pass to
loadControl API.  While this looks wrong in most cases it is very close to the same value as `getBufferedPositionUs()`.

There are two cases where it is not:

1. GAP tags (where no samples are in the "segment")
2. Cases where duration of segment is less then the media duration

Case 2 may be one source of the issue worked around by commit c19d1da6

For background, these are the rough implementations of the two methods (Both DASH and HLS):

`SequenceableLoader.getBufferedPositionUs()` &mdash; At the individual loader view this returns:

```java
  max(chunk.endTimeUs, largest timestamp in SampleQueue)
```

A the composite level, the min() of all loaders.  This allows for the case where one loader may be behind the others.

`SequenceableLoader.getNextLoadPositionUs()` &mdash; At the individual loader view this returns the `chunk.endTime`.
At the composite level, min() of all loaders