Reduces initial buffering and FF3 frame rate for trickplay
The initial buffering requirement (critera for `LoadControl.shouldStartPlayback()`)
is to prevent repeated stalling on initial playback start.   In earlier releases of trick-play this
value only considered speed = 1x, so it was inadequate for FF3 (60x).  This change increases
it but only enough to reach some threshold frame count.

Also, reduced the target frame rate for FF3 from 8 to 7 FPS, this has two parts:

1. Reduce the target frame-rate values
2. Add additional dual-mode tracks to allow a closer match from selected track to target frame-rate