Fixes first cause of WSIPCL-15765
The initial issue that is the root cause of the bug:

[WSIPCL-15765 Occasionally an attempt to use trickplay...jump to the Live point.](https://jira.xperi.com/browse/WSIPCL-15765)

The issue is that the state for the playlist updates (previous source playlist) was kept in the media playlist parser,
`FrameCuratorPlaylistParser`.  Prior to 2.15.1 the parser was created once on master playlist load so there was no issue.
With 2.15.1 update the previousPlaylist was passed to the parser as the parser was created *for each* playlist load.  What happens
then is a false `PlaylistStuckException` (because the previous playlist was lost).

Fix was simply to:

1. Pull the previous source playlist state out into the `DualModeHlsPlaylistParserFactory` so it persists for the
   whole duration of playback
2. Use the `previousMediaPlaylist` (current snapshot from the `DefaultHlsPlaylistTracker`) that is passed to the parser