Fixes stale snapshot causes false PlaylistStuckException
In 2.15.1 the support for Low-Latency HLS Playlist Delta Updates passes the current
snapshot playlist to the `HlsPlaylistParser`.   It is possible the snapshot is quite
old if the variant is left unused for a period longer then the window.  In this case the
curation logic generated the same playlist (as it saw no update), this resulted in a
`PlaylistStuckException`

To trigger this error for [WSIPCL-15765](https://jira.xperi.com/browse/WSIPCL-15765)
the steps are:

1. perform a forward trickplay
2. wait for greater then the window duration
3. perform a subsequent forward trickplay

The solution is to validate the playlist update by checking the saved source playlist
(this playlist was used to generate the possibly stale snapshot passed to the `FrameCuratorPlaylistParser`).
If the saved source is stale, simply regenerate the curated playlist as if there was
no previous snapshot.

Also, the logic that issued a warning if the curated result playlist was not valid because of
a source issue now throws the `HlsPlaylistParser.DeltaUpdateException` which causes the
`DefaultHlsPlaylistTracker` to reload the playlist immediately.