Commits HLS iFrame segment key frame only segment on EOS
Frame commit logic in the Elementary Stream Reader's depends on an AUD (Access Unit Delimiter)
or any other non-VCL NAL unit to deliniate the end of the current sample (VCL) NAL
and commit this data to the stream by calling `TrackOutput.sampleMetadata()`.
For normal HLS segments this leaves the last slice of an Access Unit un-commited
until the *next segment load* load begins and commits the first AUD NAL.

This change simply forwards the end of the file (`RESULT_END_OF_INPUT`) from the input
stream to the TS layer down to the current `ElementaryStreamReader`.  The reader will
commit the current sample, if it is a key frame sample.

Many HLS segmenters produce iFrame only segments that only include the Access Unit
with the IDR, naked of any trailing NAL's.
With this change the sample is commited as soon as on iFrame segment load completes