Bounds of EOS sampleMetadata() commit are correct
This checkin makes the H265Reader comply with "Annex B Byte stream format".

When the end of the segment bitsream is read, if the PES has a declare
length or not, and the open NAL is a key frame we commit the sample
without waiting to read the next segment (and it's AUD)

The test cases verify a pair of iFrames are read and generate the same
two committed sampleData as if they were read and commited on reading the
AUD of the next access unit.