`updateFromCurrentStats` properly allocates profile playing time
This change fixes the bug found in:

[WSIPCL-15964 - Verify the profilePlayingTime...](https://jira.xperi.com/browse/WSIPCL-15964)

The code in `PlaybackStatsExtension` that allocates playing time to
formats had an issue if:
  1. The inner states loop exists in playing state
  2. There are multiple format changes in the outer formats loop

In this case the playing time is not properly allocated to the remaining formats.

The fix handles when the *last state* is playing state(PLAYBACK_STATE_PLAYING) in `PlaybackStats.playbackStateHistory`
and one or more format changes follow. It distributes the remaining time (last time in playing state up to `toEndTime`)
to each of the formats based on their start time (`eventTime.realtimeMs`)..

Test cases added, but still in @Ignore until test case execution is fixed.