Removes liveOffsetTarget override on seek to live edge
Any calls that issue a `seekTo()` where the position is `TIME_UNSET` will
now restore the live offset target to the MediaItem set default.

The change simply passes a flag to `updatePlaybackSpeedSettingsForNewPeriod()` that
indicates the call is from a seek operation.  If this flag is set, a `positionForTargetOffsetOverrideUs`
of `TIME_UNSET` unconditionally removes the `setTargetLiveOffsetOverrideUs()`

This fixes issue #11050

Back-ported from the commit in dev-v2, functionally the same as the commit from our pull
reqeust https://github.com/tivocorp/exoplayerprvt/pull/335, so reverts that change.