Always issues `seekTo()` after trickplay exit
Refactor final seek from `switchTrickModeToNormal()` to own method, `seekOnTrickPlayExitToLastRenderedFrameN()`
This method now issues a seek unconditionally when exiting from trick play to NORMAL mode.

The final seek makes sure the `LivePlaybackSpeedContro.setTargetLiveOffsetOverrideUs()` is
always called so Live Edge adjustment is operating on the correct playback position.

This  "fixes" a bug where if we jump back to some position back in the live window then trick-play all the way to the live edge (till trickplay auto exits) the live edge adjustment was
stuck with a stale position.

Also, added the following changes:

1. updated to remove deprecated methods used (`onPositionDiscontinuity()`)
2. only empty lastRenderPositions if playing in non seek based trickplay mode
3. Added reaction time jump back feature for FF2 and greater speeds

Note this change moves the final seek till the final track selection to restore audio which is
*after* the trick-play mode change is signalled.  This has two benefits:

1. The seek happens to the non-iFrame only playlist (it is an exact seek)
2. Seek and it's associated buffering time are counted against regular playback metrics and not the trickplay operation

Lastly, removed an overly chatty log in `AdaptiveLoadControl`