Disables live adjustment on seek
The stock Google LivePlaybackSpeedControl continues to adjust the playback speed
to adjust to a live target offset no matter what the playback position is.  The
`LivePlaybackSpeedControl.setTargetLiveOffsetOverrideUs()` method is called for
every seek and, in the Google default, overrides the live offset target to the seek
position.  This "works" (keeps the live adjustment from un-doing the seek), but
continues vaccous playback speed adjustments after the seek.

This change introduces a `LivePlaybackSpeedControl` delegate that internally stops
live adjustment speed changes while the target override (seek) sets a position that
exceeds a max delta from the the true live edge.