Avoids bad seek ending trickplay at SOCU playlist start
Vecima has an issue with their iFrame only playlists where when they create a SOCU
the iFrame only playlist is behind the regular playlist, example:

$ curl -s -L "http://.../CCURStream_video_pophd0-1.m3u8?.." | head -20
  #EXT-X-PROGRAM-DATE-TIME:2023-03-30T19:59:58.859Z
  #EXTINF:6.006000,
  CCURStream_pophd0-1_T1680206398859968~D6006000.tsv?ccur_svc_type=rb

$ curl -L -s "http://.../-CCUR_iframe.m3u8?.." | head -20
  #EXT-X-I-FRAMES-ONLY
  #EXT-X-PROGRAM-DATE-TIME:2023-03-30T19:59:58.000Z

Note that `59:58.000` is 859ms behind `59:58.859`

This leads to a negative last render position on a reverse trickplay to
the start of cache, which will not play in the normal playlist.

The change works around issue https://jira.xperi.com/browse/PARTDEFECT-15432 and the
duplicates indicated

The work-around is very simple, normalize the position to 0 or greater on the final seek after transitioning.