GeekStats logs position and time delta info
Add a timed log line, method `logDebugPositionForLiveOffset()` that logs position and time
in csv format suitible for the test script to determine that multiple devices have synced on
the same position.

The log line prints once per-second, the test script simply:

1. truncates time to the nearest second to match log lines from multiple devices
2. determine the milliseconds difference in play position and time between any pair of devices
3. compute delta of the deltas to arrive at the live offset delta

```python
  delta_live = abs(device1.position - device2.position) - abs(device1.now - device2.now)
```

This can be turned on/off with `GeekStatsOverlay.setEnableLiveOffsetLogging(boolean)`, it is off
by default.