Declare foreground service type for DownloadService
This ensures the DownloadService stays functional on Android 14
where defining this type is required. On Android 14 and above,
the app also needs to define the DATA_SYNC permission, which is
added to the demo app as well. In the future, this service type
will no longer be supported and DownloadService needs to be
rewritten with another background scheduling framework.

Issue: google/ExoPlayer#11239
PiperOrigin-RevId: 548994842

----
This is a cherry pick commit for DownloadService supporting Android 14 in Exoplayer

The cherry pick commit is google/ExoPlayer@cf0ccf1, this commit was made to
dev-v2 branch and cherry-picked to releaes-v2 just after the final 2.19.1 checkin

Numerous conflicts resulted from the cherry-pick, as other fixes/changes
had been made to the DownloadService between this commit and our current
2.15.1 ExoPlayer version.

The resolution was to just take the changes in this actual commit, at least
until our version of ExoPlayer is merged up to 2.19.1

Conflicts, resolved by taking just hunks from the cherry-picked commit.
        demos/main/src/main/AndroidManifest.xml
        library/common/src/main/java/com/google/android/exoplayer2/util/Util.java