Enables HLS Live Offset Synchronization Across Multiple Devices
This commit addresses the issue of inaccurate live position measurement in HLS
by introducing NTP time synchronization. Unlike DASH, HLS lacks origin time
synchronization logic, resulting in potential discrepancies in wall-clock time
when using `System.currentTimeMillis()`.

The commit utilizes the `SntpClient` in ExoPlayer to determine the clock offset
with the default NTP server (`time.android.com`) and applies this information to the `HlsMediaSource`.

With the implementation of this change, multiple devices can now synchronize playback to
a common time source, as long as the origin server also synchronizes to an NTP time source.