Demos three modes for dual mode interaction
This commit adds support for the tenfoot demo to make a UX demos for three modes for Dual Mode
trick-play interaction using just the D-PAD.

These modes can be selected with a View intent extra, `scrub_behavior`.   The values are:

1. `scrub_only` &mdash; This mode is just a degenerate demo of SCRUB only mode, where D-PAD movements change the position
                  and key-repeats will generate some frame rate of motion while the D-PAD is held.
                  This is useful for phone UI where a touch bar is the primary input source.
2. `scrub_lp_vtp` &mdash; like `scrub_only` but as the D-PAD is held for a time it transitions to SCAN mode and
                  increases the speed the longer the D-PAD is held.   Releasing the D-PAD transitons back to SCRUB mode
3. `scrub_dpad_morph` &mdash; Basically make the D-Pad modal, entry key defines the initial mode.
                  In *Scan Mode* the D-pad acts just like the traditional TiVo series 6A remote transport D-pad.

Other controls, besides the playback position scrub bar, are included, as examples, in the playback controls overlay.  When
any of these other controls are focused the auto-hide timer is disabled.