Implements caption track check in ExoPlayer
With this change we can safely enable chunkless prepare as it performs checks
before allowing the enablement.

This change blocks enabling chunkless prepare, even if it is requested, if the
master playlist does not contain at least one explict declartion of a caption
or subtitle track (EXT-X-MEDIA ...).

While this may seem a bit draconian, most all of the live channels will have
closed captions and should have this declartion.  The fallback, ignoring the
request for chunkless prepare will not block playback.

This is decribed in the story [WSIPCL-7318](https://jira.xperi.com/browse/WSIPCL-7318)