Prevents selection of un-supported audio for sync video
The ExoPlayer `DefaultTrackSelector` does not allow modifying the mapped tracks (all the
tracks found that are supported by each renderer). The `SyncVideoTrackSelector` modifies the
`formatSupport` that is passed to it's override of `selectAudioTrack()` but it cannot modify
the `formatSupport` that it was passed.

Net result is the `MappedTrackInfo` contains `Format`s that should not be selected.

The solution is to filter these before passing them to user override track selection dialogs.
An example of how this is done is in the tenfoot demo's `ViewActivity` where it handles `KEYCODE_5`
to show an audio track selector.

Also, the `SyncVideoTrackSelector` are now defined as named constants, so you can use these for error
recovery by simply reverting back to the regular track selector and retrying playback.  Sample checks
are in ViewActivity:

```java
new PlayerErrorHandlerListener() {
              @Override
              public void playerErrorProcessed(PlaybackException error, HandlingStatus status) {
                switch (status) {
                 ...

                case FAILED:
                  switch (error.errorCode) {
                    case SyncVideoTrackSelector.ERROR_CODE_SYNC_VIDEO_FAILED_NO_PCM_AUDIO:
                    case SyncVideoTrackSelector.ERROR_CODE_SYNC_VIDEO_FAILED_TUNNELING:
                      Log.e(TAG,"Sync video failed, TODO fallback to no using SyncVideoTrackSelector", error);
                      break;
                  }
```