Enables error recovery when live sync not possible
This change adds a hook to the `SyncVideoTrackSelector` to allow it to be used all the time.
A simple enable/disable flag controls wether or not it will filter the audio tracks to ones that
support synced video.

This allows simple error recovery by just turning off the filtering and reselecting tracks, for
example:

```java
@Override
public void playerErrorProcessed(PlaybackException error, HandlingStatus status) {
  switch (status) {
    case FAILED:
      case SyncVideoTrackSelector.ERROR_CODE_SYNC_VIDEO_FAILED_NO_PCM_AUDIO:
      case SyncVideoTrackSelector.ERROR_CODE_SYNC_VIDEO_FAILED_TUNNELING:
        Log.e(TAG,"Sync video failed, restarting with filtered track selection disabled", error);
        TrackSelector trackSelector = ... // this should always be the SyncVideoTrackSelector
        if (trackSelector instanceof SyncVideoTrackSelector) {
          SyncVideoTrackSelector syncVideoTrackSelector = (SyncVideoTrackSelector) trackSelector;
          syncVideoTrackSelector.setEnableTrackFiltering(false);
          player.prepare();
        }
```

Each playback start must set the flag correctly (`SyncVideoTrackSelector.setEnableTrackFiltering()`)  before the prepare().
The flag defaults to false.