Trickplay exit now restores initial targetLiveOffset
The way `LivePlaybackSpeedControl.setTargetLiveOffsetOverrideUs()` works is
that it temporarily overrides the Live Offset set by the LiveConfiguration.  This
is so when a seek operation moves the playback position behind live edge, the
override set a new target temporarily. Once the override is over, the player sets
the override to -1, disabling live adjustment.

jIn stock Google ExoPlayer the only way to
indicate the override is over is to seek to the default live position
([Player.seekToDefaultPosition()](https://exoplayer.dev/doc/reference/com/google/android/exoplayer2/Player.html#seekToDefaultPosition())),
 and even this requires the bug fix for [this issue](https://github.com/google/ExoPlayer/issues/11050).
**NOTE** stock ExoPlayer performs Live Position adjustment even when you have
seeked far from the live edge, a feature in search of a reason to exist.

In our version of ExoPlayer we completely disable live adjustment when the position is not
within the *live range* (as defined by constant
`AugmentedLivePlaybackSpeedControl#MAX_LIVE_OFFSET_DELTA`).   This change is required because trickplay never issues a
`Player.seekToDefaultPosition()`, instead when trick-play runs out of playlist it exits the trickplay track
playback and joins the regular track at some position very close to the previous live
point.  The code in `AugmentedLivePlaybackSpeedControl.setLiveConfiguration()` is
detects trickplay has rentered the *live range*  in this way, re-enable live
adjustement and disable the override.

The bug was  the code does re-enable adjustment, but it does not release the override.