Adds NetworkLossPlayerErrorHandler to recovery choices
This change adds a new `NetworkLossPlayerErrorHandler` for seemless recovery from loss of
network that is extended enough to cause a fatal playback error.   For example unplugging the
LAN cable long enough to run out of playback buffers.

The error codes handled are `ERROR_CODE_IO_NETWORK_CONNECTION_FAILED` and `ERROR_CODE_IO_NETWORK_CONNECTION_TIMEOUT`
both of which are reported by the socket level APIs in the `HttpDataSource`.   This error
can occur for playlist load as well as segments.