Accumulates videoFramesPresented without trick-play
This change fixes [WSIPCL-18787](https://jira.xperi.com/browse/WSIPCL-18787)

The previous code depended on the fact that the `DecoderCounter`s in kept by the
ExoPlayer renderers was not reset, in fact it is each time the Render state transitions from
`STATE_DISABLED` back to `STATE_ENABLED`.   Track selection can cause this, and trick-play causes
a track selection on the video track.

The change simply captures the the `DecoderCounters` values each transiton into trick-play and
adds (`DecoderCounter.merge()`) them to a private copy object.

When you take a snapshot for reporting metrics, code in `updateValuesFromStats()` adds the current
player DecoderCounter with the captured values and return the results

When trick-play phase 2 arrives and we transition into trickplay **without** a hard
track-selection this approach will need revisiting (to use `AnalyticsListener.onVideoDisabled()`)