Implements first pass at PlaybackMetrics with ads
ExoPlayer tracks entry and exit from ad playback with events that are tracked
in the PlaybackStatsListener.  The core ExoPlayer also has knowledge of when
you are in ad playback state (Player.isPlayingAd()).

This first pass adds some logging to our playback metrics session management
and fixes any NPE's that result from ad sessions (which do not have a URL)

There will be stories to clean this up later in the One Ad project.