Hides the ad "COUNTDOWN" for trailer ads
The IMA SDK includes a number of ad UiElements, some may be configured and others
are intentionally imuteable to prevent hacking ad playback.

This change disables the Ad countdown disaplay (shows "(Ad N of M 0:15)") when playing
HPT (Trailer Ad) but leaves it on for regular ad type.

Also, added method `setHideAdControlsForTrailer()`, this setting makes the entire ad container
invisible for trailer playback.  This is a pretty invasive and brute force hack
option to eliminate undesireable UI elements, the default is off.

The prefered method is authoring the VAST document to control it, eg:

1. *skipButton* – if the ad can be skiped, only shown if the <Linear> element has a skipoffset time
   attribute.
2. *adClick* – the click action (shown as focus rectangle around the entire ad video),
   it simply fires the TAPPED event, with no other action.  It will probably not be possible
   to focus in Hydra
3. *wta* (Why This Ad) – this is Google’s extension that explains and allows customization
   of the user tracking. Excluding the Icon from the generated VAST and it will not be shown in
   the UI (see https://support.google.com/admanager/answer/10678356?hl=en#wta)