Final sample is commited on EOS for MPEG-TS
For Transport Stream packaged Elementery Streams the exact length of a NAL are
not always known.  This requires using the algorithm in *Rec. ITU-T H.264 Annex B* to determine i
the length of a NALU.  In addition, Transport Stream packaging lacks the sample table, provided 
by fragmented MP4 encoding, to indicate the size of a sample frame.

The  H.264/265 readers follow *Section 7.4.1.2.3 Order of NAL units* of *Rec. ITU-T H.264* 
to determine when to end of a sample frame (coded picture) is and call
`TrackOutput.sampleMetadata()`.   They use the AUD (NAL type 9), that is the start of the 
**next sample frame** in order to commit the sample frame.

While this algorithm works for a single MPEG-TS file, for HLS it leaves an orphan frame
at the end of each HLS segment, until the next segment is read.   Additionally, the last frame of the
last HLS segment is dropped completely.

The solution is simple, treat the end of the HLS segment file as the end of a sample frame. 
HLS segmenters will never split a frame across segments.

Fixes xperi bug: PARTDEFECT-20827 [Second I-Frame
(sample) always skipped upon FF activation]

The following UTs have been added along with Ateme generated samples:
- consumeThreeH264IFrames
- consumeThreeH265IFrames

Review comments #1: removing zero byte injection

Expanding sample release logic for regular video segments