Fix bug where start-over ends early for DASH
This commit fixes the SOCU issue, as reported to AndroidX/media:
https://github.com/androidx/media/issues/1441

Our matching bug is [WSIPCL-19084](https://jira.xperi.com/browse/WSIPCL-19084)

The fix is simply to determine that an existing DASH `Period` has been updated
in the transition from dynamic to static and triggers a seek to the new period position
This position should be the exact same media position, however because of the
fact that ExoPlayer normalizes everything to *Period Time* as segments are scheduled
to download (based on the `SegmentTemplate@presentationTimeOffset`) the change
to `presentationTimeOffset` is not reflected in the existing queued segments so they
must be flushed with a seek.

Prelimiary testing with both DASH and HLS SoCu works with this change (HLS SoCu will not
tigger the seek) and only a change from dynamic to static playlist should fire this logic.
Still this is a very complex central area of ExoPlayer so we will watch the pull request
https://github.com/androidx/media/pull/1451 carefully for a final change.

Along with the change, fixed up the GeekStats display of posiiton to show HH:MM:SS.ms for
VOD and removed excess logging in the timebar.