Only commit sample on EOS if it is an i-frame
This change updates the processing for both  `H264Reader.endOfStream()` and
`H265Reader.endOfStream()` so they only process samples that are considered
key frames at the end of stream.  This prevents the existing Google test cases from failing.

The Google test cases expect the last sample of a frame is not commited until the next
segment.   They have a fix for this in Androidx, so we can revisit the change then.

Note, both our changes and the existing Google decoding releases the H.264
elementary streams for a frame slightly differently for the same TS segment data.

Using ffmpeg:

```
ffmpeg segment.ts -an  -c:v copy -vbsf h264_mp4toannexb output.h264
```

Gives:

```
$ od -Ad -tx1 output.h264 | head -1
0000000    00  00  00  01  09  10  00 ...
$ od -Ad -tx1 output.h264 | tail -2
0019008    50  7c  28  d5  bb  c6  d3  82  ce  d3  de  44
```

But the H26xReaders both produced frames where the leading `zero_byte` is stripped and
instead included at the end of the previous frame.  With this latest change the trailing 0 byte is not included.

Something to check for when we update to the new Google code in Androidx but for now this should be
of no matter to the decoders