Supports hook altering properties post  `MediaDrm` creation
There is at least one use-case where the platform default properties for the `MediaDrm`
are not the desired set.  This change adds a hook to allow access to the `MediaDrm` objects
created by `DefaultDrmSessionManagerProvider`

One use-case is mosaic views.   Typically these are lower resolution playback that can be
satisfied by only Widevine L3 security.  If the platform default is L1 (as most are now) the
platform is unlikely to support more than 2 simultaneous streams in the hardware TEE.

An example to support this would be:

````java
sessionManagerProvider.setExoMediaDrmProvider(uuid -> {
  try {
    FrameworkMediaDrm mediaDrm = FrameworkMediaDrm.newInstance(uuid);
    if (WIDEVINE_UUID.equals(uuid)) {
      mediaDrm.setPropertyString("securityLevel", "L3");
    }
    return mediaDrm;
  } catch (UnsupportedDrmException e) {
    return new DummyExoMediaDrm();
  }
````