Retry AudioTrack with brief delay. Fixes PARTDEFECT-21565
Broadcom is using separate decoders for AC3 stereo and 6 channel.

When the stream changes from 6 to 2 channel AC3, the AudioTrack becomes NULL because Broadcom switched to another audio decoder.
In such case DefaultAudioSink needs to create a new AudioTrack. But creating AudioTrack immediately leads to an exception because
Android is still in stop() routine. So, Broadcom wants us to wait for about 200 msec before calling createAudioTrack.

ExoPlayer did similar fix in -

1. [82ee826](https://github.com/google/ExoPlayer/commit/82ee82635be189743eac82c8e2269f0cfd40858e) for google/ExoPlayer#8696
2. [82ee826](https://github.com/google/ExoPlayer/commit/82ee82635be189743eac82c8e2269f0cfd40858e) for #9712

Both these commits went to ExoPlayer 2.17.0. We will get them along with 2.19.1 update.
This is not a direct cherry-pick as 82ee826 wont fix our issue. Hence making custom changes to address this tunneling mode issue.