Align forceAsyncQueueingSynchronizationWorkaround setting for SEI platforms in demo app with Hydra app
This change disables the forceAsyncQueueingSynchronizationWorkaround for SEI Robotics platforms in the demo app, bringing
it in line with the settings used in the full Hydra app.

The workaround is now disabled for both Amazon and SEI platforms to prevent unnecessary
blocking and improve performance consistency in the demo for Multiview evaluation of SEI based platforms.