Works around issue with DASH ad playback for pre-roll
The TiVo issue is [WSIPCL-21072](https://jira.xperi.com/browse/WSIPCL-21072)

GAM seems to generate VAST documents with media that does not
have an extension to indicate the media type on the URL, this confuses ExoPlayer.
This occurs for DASH MPD's (this is what we have seen so far).

The AndroidX/Media and IMA SDK issues is opened [AndroidX/Media#2039](https://github.com/androidx/media/issues/2039)

This commit works around the issue by limiting the supported media to exclude DASH.