Adds audio focus support the the `MultiExoPlayerView`
This commit adds an `MultiPlayerAudioFocusManager` that the `MultiExoPlayerView` uses to obtain  audio focus prior to starting
playback.

When the `MultiViewPlayerController.playMediaItem()` method is called
for the `MultiViewPlayerController` that is selected in the grid the
`MultiPlayerAudioFocusManager` requests focus (if the application does not
already have focus), playback starts for all non-selected players and in paused mode
for the selected player until audio focus is obtained.

The two methods in `MultiExoPlayerView` either stop the players, keeping focus or
completely release the players:

1. `MultiExoPlayerView.removeAllPlayerViews()` -- this method relenquishes audio
focus completely, calls `release()` on all the players, destroy's the `PlayerView`'s and
frees all resources.   This should be called when the main activity is stopped (onStop())

2. `MultiExoPlayerView.stopAllPlayerViews()` -- this method simply calls `stop()` and `clearMediaItems()` on all the players to stop playback and free any buffered media.
This method can be called to simply quiece the multi-view when switching to a regular single
player view.

Related small changes:
- Fixed doc comment in `SimpleExoPlayerFactory.setRendererState()`
- Added method to disable audio rendering to `SimpleExoPlayerFactory`
- Made members package private that should not be called from clients