Optimizes selection of active multi-view cell
With this change, selecting a different cell for audio takes milliseconds instead of the 2-3
seconds required to enable/disale the renderer.

This is an optional setting that prioritzes audio formats that permit mute/unmute in the
track selection algorithm, falling back gracefully if that is not possible.

These audio formats provide a few features useful for multi-view:

1. They are substantially lower bitrate  (for example, AAC is at the most 1/2 the bitrate of AC3)
2. As they are converted to PCM audio for input to the audio track AAC allows volume setting as
   well as speed adjustment to facilitate multi-view live edge sync.

The cost is a small additional overhead of downloading audio for the non-playing cells.