Chunkless-prepare creates an additional ID3 track. Thats done in commit 400619C for the issue#4016
The ID3 track now adds an additional track to trackgroupcount. But when there is no ID3 track declared in the manifest, there is no SampleQueue created for ID3 in HlsSampleStreamWrapper

That leads to an index with value -1 in trackGroupToSampleQueueIndex. Assertion is hit when the sampleQueuesEnabledStates array points to -1

The ID3 tracks is added to optionalTrackGroups. Functions unbindSampleQueue() and selectTracks() should check for optionalTrackGroups before throwing an exception

If the trackIndex is present in optionalTrackGroups, treat that as non_fatal error