Restores volume to un-muted when selected un-conditionally.
This fixes [WSIPCL-21724](https://jira.xperi.com/browse/WSIPCL-21724)
"[MULTIVIEW] Fix AAC audio start on subsequent multiview entry"

The issue occurs because it is not known if the audio Format supports volume mute
prior to the initial track selection.  This will cause the code to fail to restore the
volume to un-muted level initially.

This fix is quite simple, always restore the volume mute/un-mute state on selection.
Then only leave the audio Renderer enabled if it is known the audio Format supports
volume mute/un-mute.