Adds a specific layout for single row MultiView
**weighted layouts**  require **multiple measurement passes** to determine the
correct size distribution among child views.  This presents a problem for the `SurfaceView`
as no hardware surface will be created for a 0x0 view, and thus the first frame will not render.

The `PlayerView` layout includes a `SurfaceView` for the player.   The player surface is nested in an `AspectRatioFrameLayout`, which centers and letterboxes the player surface.

This combination presents a chicken and the egg problem for establishing proper sizes for a
GridLayout when the weightings are other that 1.0 (even distribution)

The solution here is to use a simpler layout with margins for single row (1x2) and reserve
using the GridLayout for more than 2 cells.

Both layouts are merge inflated into the `MultiExoPlayerView` ViewGroup, as they share a
common top container layout class (`LinearLayout`).  With this any `LinearLayout`
layout attibutes can be added to the XML layout tag that creates the `MultiPlayerView`

Other fine-tuning to the layout / behavior:

1. Added remote key to the `MultiViewActivity` to switch views on the fly (from 2 to 4
   cells).  The number 2 and number 4 keys do this.
2. Moved the visible selection treatment to a "border" drawn around a frame containing the
   Player's aspect ratio sized contianer that fills the parent, so visually the focus rectangles
   are all the same size

NOTE: next checkin will cleanup refactor rename the layout XML files so it makes more sense.
not done on this checkin so the diffs are more clear.