`MultiViewPlayerController` create distinct `DefaultBandwidthMeter`s
The default behavior for ExoPlayer, when using the `SimpleExoPlayer.Builder`, is to
share the same singleton `DefaultBandwidthMeter`:

```java
public Builder(
    Context context, RenderersFactory renderersFactory, ExtractorsFactory extractorsFactory) {
  this(
      context,
...
      DefaultBandwidthMeter.getSingletonInstance(context),
```

This has a benefit where one Builder is used to create multiple ExoPlayer's in a serial fashion,
as they can share the current bandwidth estimate from any previous playback.

With multi-view the players are running concurrently, all feeding their downloads into this
singleton `BandwidthMeter`.   This produces a lot of duplicate measurements.